cambio de int a long para seguir con el mismo tipo de la clase de los profes

+++ EjerciciosAplicacion2/Ejercicio2EquipoB.java
@@ -23,8 +23,8 @@ public class Ejercicio2EquipoB {
         }
     }
 
-    public static int sumaHastaPrimo(int n) {
-        int suma = 0;
+    public static long sumaHastaPrimo(long n) {
+        long suma = 0;
         int contador = 0;
         if (isPrime(n)) {
